+++ src/frontend.ts
@@ -1921,6 +1921,7 @@ export class Frontend extends EventEmitter<FrontendEvents> {
               } else if (data.params.value === 'Fatal') {
                 Logger.level = MatterLogLevel.FATAL;
               }
+              this.matterbridge.matterLogLevel = MatterLogLevel.names[Logger.level as number] as LogLevel;
 
               // Set the global logger callback for the WebSocketServer to the common minimum logLevel
               let callbackLogLevel = LogLevel.NOTICE;